Split KafkaConnectBuildAssemblyOperatorTest.java
